fix python3 true divide

+++ rcnn/symbol/proposal_target.py
@@ -27,8 +27,8 @@ class ProposalTargetOperator(mx.operator.CustomOp):
     def forward(self, is_train, req, in_data, out_data, aux):
         assert self._batch_rois % self._batch_images == 0, \
             'BATCHIMAGES {} must devide BATCH_ROIS {}'.format(self._batch_images, self._batch_rois)
-        rois_per_image = self._batch_rois / self._batch_images
-        fg_rois_per_image = np.round(self._fg_fraction * rois_per_image).astype(int)
+        rois_per_image = int(self._batch_rois / self._batch_images)
+        fg_rois_per_image = int(round(self._fg_fraction * rois_per_image))
 
         all_rois = in_data[0].asnumpy()
         gt_boxes = in_data[1].asnumpy()
